fix: make the build output of the component regain its type
