+++ src/runtime/components/css.ts
@@ -3,7 +3,8 @@ import { computed, watch, h, defineComponent } from 'vue'
 import { getIconCSS } from '@iconify/utils/lib/css/icon'
 import type { PropType } from 'vue'
 import type { IconifyIcon } from '@iconify/types'
-import type { NuxtIconRuntimeOptions, NuxtIconRuntimeServerOptions, IconifyIconCustomizeCallback } from '../../types'
+import type { NuxtIconRuntimeOptions, NuxtIconRuntimeServerOptions } from '../../types'
+import type { IconifyIconCustomizeCallback } from './shared'
 import { loadIcon, resolveCustomizeFn } from './shared'
 import { useAppConfig, useNuxtApp, useHead, useRuntimeConfig, onServerPrefetch } from '#imports'
